+++ packages/tests/services/load.spec.js
@@ -1,8 +1,0 @@
-import { useLoad } from '@studiometa/js-toolkit';
-
-describe('The `useLoad` service', () => {
-  it('sould return its props', () => {
-    const load = useLoad();
-    expect(Object.keys(load.props())).toEqual(['time']);
-  });
-});